[IMP] make price_unit_manual a non-computed field This commit changes price_unit_manual from a stored computed field to a non-computed field.
This is due to the behavior described in https://github.com/odoo/odoo/issues/14279
- when "Send by Email" is pressed in SO, Odoo re-calculates computed fields for
report generation, and price_unit_manual would always result to be zero, which is
different to the DB record value, and the price presentation becomes incorrect
in generated report. This commit is to work around this problematic behavior.